+++ packages/react-hook/src/use-debounce-state.ts
@@ -1,5 +1,5 @@
 import { useEffect, useMemo, useState } from 'react'
-import { debounce } from '@starkit/utils'
+import { debounce } from '@starkit/utils/debounce'
 
 export function useDebounceState<T>(value: T, delay = 1000) {
   const [debouncedValue, setDebouncedValue] = useState(value)