+++ packages/react-hook/tsup.config.ts
@@ -1,0 +1,10 @@
+import { globSync } from 'glob'
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  entry: globSync('src/*.ts'),
+  format: ['esm', 'cjs'],
+  dts: true,
+  clean: true,
+  external: ['react']
+})