+++ packages/utils/src/error.ts
@@ -1,12 +1,15 @@
 export class AbortError extends Error {
   constructor(opt?: ErrorOptions) {
-    super('The operation was aborted', opt)
+    super('This operation was aborted', opt)
   }
   get name() {
     return 'AbortError'
   }
   get cancelled() {
     return true
+  }
+  get code() {
+    return 20
   }
 }
 
@@ -16,6 +19,9 @@ export class TimeoutError extends Error {
   }
   get name() {
     return 'TimeoutError'
+  }
+  get code() {
+    return 23
   }
 }
 
@@ -35,4 +41,16 @@ export function isUnAuthenticated(
   err: unknown
 ): err is StatusError & { status: 401 } {
   return err instanceof StatusError && err.status === 401
+}
+
+export function isAbortError(
+  err: unknown
+): err is Error & { name: 'AbortError' } {
+  return err instanceof Error && err.name === 'AbortError'
+}
+
+export function isTimeoutError(
+  err: unknown
+): err is Error & { name: 'TimeoutError' } {
+  return err instanceof Error && err.name === 'TimeoutError'
 }