+++ packages/event/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @starkit/event
 
+## 1.0.6
+
+### Patch Changes
+
+- support a better tree shaking imports
+- Updated dependencies
+  - @starkit/lifecycle@1.0.3
+  - @starkit/utils@1.0.2
+
 ## 1.0.5
 
 ### Patch Changes